Phase 3: Complete enhanced server integration
- Replaced original server with enhanced version using connection pooling
- Integrated all 7 analysis tools with standardized JSON responses
- Added hybrid parameter support (MCP config + manual parameters)
- Switched to stdio transport for Q Chat integration (with HTTP fallback)
- Enhanced connection management with automatic pool management
- Added comprehensive error handling with actionable suggestions
- Maintained backward compatibility for existing tool signatures
- Added test suite to verify all components work correctly
- Ready for production use with Q Chat and standalone deployment
